Small fixes. Counters are now a long int. Linux compatibility. psi4 code outputs separate energies.

+++ config.h
@@ -19,7 +19,7 @@ struct configuration_t
 
 	/* "sampling" section */
 
-	int iterations;
+	long int iterations;
 	int thermalization;
 	double timelimit;
 	int decorrelation;
